+++ Intranet.Web/ClientApp/src/components/Header.vue
@@ -1,8 +1,9 @@
 <template>
   <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
     <router-link to="/" class="navbar-brand">Home</router-link>
-    <router-link to="admin" class="navbar-brand">Admin</router-link>
-  
+    <router-link v-if="isAuthenticated" to="admin" class="navbar-brand">Admin</router-link>
+    
+    
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
       <ul class="navbar-nav mr-auto">
         <li class="nav-item active">
@@ -26,10 +27,6 @@
           <a class="nav-link disabled" href="#" tabindex="-1" aria-disabled="true">Disabled</a>
         </li>
       </ul>
-      <!-- <form class="form-inline my-2 my-lg-0">
-        <input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search">
-        <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
-      </form> -->
       <sign-in />
     </div>
   </nav>
@@ -37,11 +34,24 @@
 
 <script>
 import SignIn from './SignIn.vue';
+import { mapState } from 'vuex';
+import IdentityService from '@/services/identity.service';
 
 export default {
   name: 'Header',
+  data() {
+    return {
+    }
+  },
   components: {
     SignIn
+  },
+  created() {
+  },
+  computed: {
+    ...mapState({
+      isAuthenticated: state => state.auth.isAuthenticated
+    })
   }
 }
 </script>
