+++ Intranet.Web/ClientApp/src/views/Admin.vue
@@ -1,14 +1,21 @@
 <template>
-  <div class="container-fluid">
-    Welcome to the Administrator Panel!
-    <br/>
-    <br/>
-    <user-prop-edit />
+<div class="container-fluid">
+    <div v-if="isAuthenticated">
+      Welcome to the Administrator Panel!
+      <br/>
+      <br/>
+      <user-prop-edit />
+    </div>
+    <div v-else>
+      Sorry you must be logged to use Administrator Panel!
+    </div>
   </div>
 </template>
 
 <script>
-import UserPropEdit from './../components/UserPropEdit.vue'
+import UserPropEdit from './../components/Admin/UserPropEdit.vue'
+import IdentityService from '@/services/identity.service';
+
 export default {
   name: 'app',
   data() {
@@ -18,6 +25,9 @@ export default {
   },
   components: {
     UserPropEdit
+  },
+  computed: {
+    isAuthenticated : IdentityService.isAuthenticated
   }
 }
 </script>
